ai_backend/requirements.txt
langgraph
langchain
langchain-openai
langchain-community
pydantic==2.8.2
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.30.6
numpy==1.26.4
scipy==1.14.0
matplotlib==3.8.4
sympy==1.13.2
reportlab==4.2.2
PyPDF2==3.0.1
python-multipart==0.0.9
httpx==0.27.0
tiktoken==0.7.0
scikit-learn==1.5.1
nltk==3.8.1
